fix smile interface functino change

+++ Codenite/Assets/Script/PlayerManager/UI/SmileBar.cs
@@ -17,7 +17,7 @@ public class SmileBar : MonoBehaviour
         fill.color = gradient.Evaluate(1f);
     }
 
-    public void SetValue(int value)
+    public void SetValue(float value)
     {
         slider.value = value;
         fill.color = gradient.Evaluate(slider.normalizedValue);
